Change methods and data away from the jargon "ADC", to "Sample".
It is admittedly bad practice to change the definition of a data class,
particularly, without changing the ClassDef version, but that is what this
commit does.  The reason for doing it now is that this class has not been
used for any actual data production, and is not part of any stable freeze.