Remove THit derived classes that will not be used.
These are old THit derived classes that were not found to be useful for the
ND280.  They mostly duplicate functionality in TReconHit.  The TMultiHit
object is the most useful since it can be used to replicate a FADC signal,
but in ND280 we found that the FADC pulse could be easily processed at the
digit level and the hits could just represent the final pulse positions,
times, &c.  In anycase, the current TMultiHit object was not optimal for
the problem it tried to solve.