Check rhs for null ptr in the "==" operator.
Make sure that the lhs and rhs both have valid handle pointers before
getting the object and checking for equality.  In fact, for any valid
handle, the fHandle field will not be NULL, so this is an over kill.  This
will come into play when dealing with a "NULL" handle, which should be
protected against explicitly in user code.  However, this code should also
be safe.