Redefine track width and show cone angle.
The original TReconState definitions had two components each for the track
width and shower cone opening angle in the TCorrValue state.  This makes
since geometrically since the two components corresponded to the two
directions perpendicular to the track direction, but most of
our expected objects are nominally cylinderically symmetric.  More
concretely, the two perpendicular directions have never been uniquely
defined, and it is "difficult" to make a simple, easy definition without
adding extra information to the track state.  Given the practical and usage
considerations, I'm imposing the constraint that the track and showers have
a (local) cylinderical symmetry.