Change GetHitSelection to GetHits
There was previously two naming conventions for accessing the hit
selection in an object (GetHitSelection and GetHits) that were not used
uniformly among the classes.  The GetHits() method name was used in
reconstruction objects, and the GetHitSelection() name was used in the
event and algorithm result.  This made it more difficult to remember which
method went with which class.  This commit makes the naming more uniform so
that all of the hit access methods are named "GetHits()", and they all have
the same fingerprint.