Clean up the integer type definitions in TEventContext.
This updates the TEventContext to only use unsigned integers and to specify
them in a machine dependent way (using the ROOT UInt_t definitions).  This
also extends the method interface to document the "spill" field as holding
either the spill number or number of nanoseconds since the last second
tick.  The convenience methods <Get/Set>Nanosecond() are added.