Add fields to track the amount of ionization deposit.
The energy deposit per segment is now track as "total" and "secondary".
This is provided so that ND280MC has the tools to separately track the
amount of ionization and the amount of scintillation.  The plan is that
total is the total continuous energy deposit, the secondary is the amount
that goes into ionization, and total minus secondary is the amount that
goes into scintillation.