Save all of the charge samples associated with a hit.
This adds the interface to return the start and stop time of a hit, and all of the charge samples associated with the hit.  The default in THit is to assume that the hit has one sample.  This is overridden in TSingleHit where all of the charge samples are kept in a vector.  This can be set in TWritableDataHit.