Implement memory management for weak THandle objects.
This implements separate counters for the objects referenced by the
THandle, and the THandleBase objects used to reference count the objects.
This allows classic reference counted weak pointers to be implemented.  The
object is deleted if the reference counter goes to zero, and the
THandleBase object (plus the object) is deleted if the handle counter goes
to zero.  The user interface to make weak handles is not implemented in
this commit.